+++ apsw/tests.py
@@ -3473,8 +3473,6 @@ class APSW(unittest.TestCase):
         check("I", "en_us", equal=True)
 
     def testJSON1Extension(self):
-        if not self.checkOptionalExtension("json1", "select json('{}')"):
-            return
         # some sanity checks that it is working
         l = self.db.cursor().execute("select json_array_length('[1,2,3,4]')").fetchall()[0][0]
         self.assertEqual(l, 4)
@@ -5387,8 +5385,7 @@ class APSW(unittest.TestCase):
                     f"file { filename } function { name } calls PyGILState_Ensure but does not have MakeExistingException"
                 )
         # not further checked
-        if name.split("_")[0] in ("ZeroBlobBind", "APSWVFS", "APSWVFSFile", "APSWBuffer", "FunctionCBInfo",
-                                  "apswurifilename", "APSWFTS5Tokenizer"):
+        if name.split("_")[0] in ("ZeroBlobBind", "APSWVFS", "APSWVFSFile", "APSWBuffer", "FunctionCBInfo", "APSWFTS5Tokenizer"):
             return
 
         checks = {
@@ -5470,6 +5467,11 @@ class APSW(unittest.TestCase):
             "apswfcntl": {
                 "req": {}
             },
+            "apswurifilename": {
+                "req": {
+                    "check": "CHECK_SCOPE"
+                }
+            }
         }
 
         prefix, base = name.split("_", 1)
@@ -10413,12 +10415,22 @@ SELECT group_concat(rtrim(t),x'0a') FROM a;
 
             def xOpen(self, name, flags):
                 assert isinstance(name, apsw.URIFilename)
-                name_catch.append(name)
+                name_catch.append((name, str(name)))
                 raise apsw.SQLError()
+
+            # MacOS fails the name we provide returning
+            # cantopen for this, so we avoid their code
+            def xFullPathname(self, name: str) -> str:
+                return name
+
         t = TVFS()
 
         with contextlib.suppress(apsw.SQLError):
-            apsw.Connection("/tmp/uri_test", vfs="uritest")
+            with tempfile.NamedTemporaryFile() as n:
+                apsw.Connection(n.name, vfs="uritest")
+
+        self.assertEqual(len(name_catch), 1)
+        uriname, urinamestr = name_catch[0]
 
         objects = (self.db,
                    db2,
@@ -10429,7 +10441,7 @@ SELECT group_concat(rtrim(t),x'0a') FROM a;
                    apsw.zeroblob(3),
                    blob,
                    blob2,
-                   name_catch[0],
+                   uriname,
                    apsw.VFS("aname", ""),
                    apsw.VFSFile("", self.db.db_filename("main"),
                        [apsw.SQLITE_OPEN_MAIN_DB | apsw.SQLITE_OPEN_CREATE | apsw.SQLITE_OPEN_READWRITE, 0]),
@@ -10438,6 +10450,27 @@ SELECT group_concat(rtrim(t),x'0a') FROM a;
         )
         for o in objects:
             self.assertNotEqual(repr(o), str(o))
+            # issue 501
+            if isinstance(o, apsw.URIFilename):
+                self.assertNotEqual(repr(o), urinamestr)
+                self.assertNotEqual(str(o), urinamestr)
+
+        # more issue 501
+        with contextlib.suppress(ValueError):
+            uriname.filename()
+            1/0
+        with contextlib.suppress(ValueError):
+            uriname.parameters
+            1/0
+        with contextlib.suppress(ValueError):
+            uriname.uri_boolean("name", False)
+            1/0
+        with contextlib.suppress(ValueError):
+            uriname.uri_int("name", 0)
+            1/0
+        with contextlib.suppress(ValueError):
+            uriname.uri_parameter("name")
+            1/0
 
     # This test is run last by deliberate name choice.  If it did
     # uncover any bugs there isn't much that can be done to turn the