+++ doc/changes.rst
@@ -20,10 +20,14 @@ Added :func:`apsw.ext.result_string` to turn an result code into
 a string, taking into account if it is extended or not.
 
 Provide detail when C implemented objects are printed. For example
-:class:`connections <Connection>` includes the filename.
+:class:`connections <Connection>` include the filename.
 (:issue:`494`)
 
 Added :meth:`URIFilename.parameters` (:issue:`496`)
+
+:class:`URIFilename` are only valid for the duration of the
+:meth:`VFS.xOpen` call.  If you save and use the object later you will
+get an exception.  (:issue:`501`)
 
 3.44.0.0
 ========